完善系统的异常功能，修改BUG

+++ code/activitiOA/src/main/java/com/cypher/activiti/model/User.java
@@ -8,7 +8,7 @@ public class User implements java.io.Serializable {
 
 	private Long userId;
 
-	private Integer deptId;
+	private Long deptId;
 
 	private String loginName;
 
@@ -40,11 +40,11 @@ public class User implements java.io.Serializable {
 		this.userId = userId;
 	}
 
-	public Integer getDeptId() {
+	public Long getDeptId() {
 		return deptId;
 	}
 
-	public void setDeptId(Integer deptId) {
+	public void setDeptId(Long deptId) {
 		this.deptId = deptId;
 	}
